pKVM: VMX: Handle EXIT_REASON_MSR_READ
Handle EXIT_REASON_MSR_READ cooperatively by the pKVM and the host. The
pKVM will try to emulate this MSR first. If such MSR is not emulated by
the pKVM and allowed to be emulated by the host, the pKVM hypervisor
will return back to the host from the vcpu_run PV interface, and the
host will emulate this via the kvm_emulate_rdmsr().

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>