pKVM: VMX: Handle EXIT_REASON_BUS_LOCK
Enable the bus_lock detection in the pKVM hypervisor if this is
requested by the host. The EXIT_REASON_BUS_LOCK is handled by the host.
The pKVM hypervisor returns back to the host from the vcpu_run PV
interface and the host handles this exit similarly as the native KVM
host does, which deliver this to the user space VMM to handle.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>