KVM: pVMX: Implement get_mt_mask operation
As the pKVM VMX host also uses the EPT which is managed by the host
itself as the guest stage-2 MMU, the memory type mask bits are the same
for pKVM VMX host and the native KVM VMX host. So reuse the
vmx_get_mt_mask() for the get_mt_mask operation.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>