pKVM: x86: Save the host state for xsave switching
When load a guest vCPU, save the host xsave state, i.e., PKRU and
IA32_XSS so that later when running this guest vCPU, the up-to-date
state can be switched between the guest and the host. XCR0 is also
switched but it is not necessary to be saved when load a guest vCPU as
it can be updated when handling the xsetbv vmexit from the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>