KVM: pVMX: Add kvm_x86_init_ops instance
Add kvm_x86_init_ops instance for the host. This will be used by the KVM
core to initialize the vendor operations. The instance is initialized
with the pkvm_host_vt_x86_ops for the runtime_ops, and stub code for the
pmu_ops as the guest PMU is not supported.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>