pKVM: x86: trace: Add host vmexit stats tracing support
Add support to trace each host vmexit on each CPU for the profiling
purpose. Each host vcpu will have a struct vmexit_perf which is used to
record the stats for each vmexit reason, i.e., the number of exits and
the cycles used to handle these exits.

By default, the tracing is turned off. It can be turned on by the host
via a specific PV interface which is implemented by the following patch.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>