pKVM: x86: trace: Add enable_vmexit_trace PV interface
The enable_vmexit_trace PV interface can be used by the host to turn on
or turn off the vmexit tracing in the pKVM hypervisor. Once the tracing
is turned on, the pKVM hypervisor will start to record the vmexit stats
for each host vmexit. And when tracing is turned off, no vmexit stats
will be recorded any more. Re-starting the tracing will result in the
pKVM hypervisor to clear the previous recorded vmexit stats.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>