pKVM: x86: Enforce initial values of most MSRs
Even though most MSRs are emulated by pKVM itself, not by the host,
pKVM still allows the host to set initial values of any MSRs for pVMs,
which is a potential source of security issues.

So implement a simple policy similar to the one we implemented for
initial CPUID values: allow the host to pre-initialize only those
MSRs that are currently pre-initialized by crosvm.

Additionally allow the host to set any MSR to the same value it is
already set to by pKVM (in such case there's nothing to do anyway, so
just return success to the host instead of failure). This addresses
cases like KVM-high setting MSR_IA32_XSS to 0 in kvm_vcpu_reset() or
crosvm setting MSR_STAR, MSR_CSTAR etc to 0.

MSRs that are managed by the host (e.g. APIC, TSC) are not affected,
since the host doesn't need to ask pKVM to set them anyway.

Signed-off-by: Dmytro Maluka <dmaluka@google.com>