pKVM: VMX: Improve the host MSR vmexit handler
With the VMX MSR bitmap support, the MSR which the corresponding bit in
the bitmap is 1 will cause the vmexit to the pKVM. So if an MSR in the
bitmap, it should be emulated otherwise it is a code bug.

If an MSR out of the range of 00000000 ~ 00001FFFF and C00000000 ~
C0001FFF, accessing will cause vmexit unconditionally which was handled
by injecting #GP to the host. However such MSR may be implemented by the
platform. So to make sure these MSR can be emulated properly, use the
safe version MSR accessing wrapper to access this MSR first. If it is
supported, pass through the accessing result to the host. Otherwise
inject the #GP exception.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>