SQUASHME: pKVM: x86: Avoid using host's .data..once in exception fixup
As a part of reusing the code from arch/x86/mm/extable.c to support
exception fixup in pKVM hypervisor, we are reusing the __ONCE_LITE_IF()
macro, in order to print an error when reading or writing an
unsupported MSR first time at the given place in the code. As a result,
since __ONCE_LITE_IF() is using the host kernel's .data..once section,
we are inadvertently using this data section in the hypervisor as well.
That is ok to do in pKVM debug mode which does such things anyway (and
which is hacky and insecure anyway), but in normal mode we should avoid
relying on using host kernel's data in pKVM hypervisor at runtime. [*]

Luckily it is only used for printing errors, which only works in the
debug mode anyway. So we can fix this just by not using __ONCE_LITE_IF
in non-debug mode.

[*] Although in this particular case we may not have any real security
    issue here. Nevertheless, this change may still be useful, for
    instance, if in the future we rework pKVM as follows: instead of
    mapping the entire host memory in the hypervisor page tables, only
    map those parts that are actually used by the hypervisor (similarly
    to pKVM-ARM), as that would be useful in general, for easier
    detection of accidental bugs and security issues in pKVM code.

Fixes: f0d4bc4337b7 ("pKVM: x86: Support exception fixup")
Signed-off-by: Dmytro Maluka <dmaluka@google.com>