pKVM: x86: Sort the exception table exception
Sort the exception table section so that to make the exception table
searching faster with the binary search.

The exception table is in the pKVM's rodata section. Once switch to the
pKVM's MMU, the rodata section is mapped with read-only thus sorting the
exception table cannot be done as it requires read-write permission. So
sorting the exception table with the host MMU which still maps the pKVM's
rodata section with read-write permission before switching to the pKVM's
MMU.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>