pKVM: x86: Import the kernel find_bit and hweight libraries
Add support for bit operations in the pKVM hypervisor by importing the
kernel's find_bit and hweight libraries. The find_random_bit() is
re-implemented as an inline function to cause a build bug as this
function is not used by the pKVM hypervisor and the original
implementation requires external symbols which are not able to be
imported. So just causing a build bug to detect if it is used by the
pKVM hypervisor or not at the compiling time.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>