pKVM: x86: Add set/get_rflags PV interfaces
Add set/get_rflags PV interfaces to allow the host to access guest
rflags via the vendor-specific set/get_rflags operations. For a pVM,
these two PV interfaces allow the host to access the rflags before the
pVM has started running, but doesn't allow to access it if the pVM is
already running.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>